fix: separate logic
